+++ mobile/android/android-components/docs/changelog.md
@@ -22,6 +22,15 @@ permalink: /changelog/
 * **support-ktx**
   * Added property delegates to work with `SharedPreferences`.
 
+* **feature-customtabs**
+  * Added `CustomTabsServiceStore` to track custom tab data in `AbstractCustomTabsService`.
+
+* **feature-pwa**
+  * Added support for hiding the toolbar in a Trusted Web Activity.
+  * Added `TrustedWebActivityIntentProcessor` to process TWA intents.
+  * Added `CustomTabState.trustedOrigins` extension method to turn the verification state of a custom tab into a list of origins.
+  * Added `WebAppHideToolbarFeature.onTrustedScopesChange` to change the trusted scopes after the feature is created.
+
 # 10.0.1
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v10.0.0...v10.0.1)